requirements.txt
certifi==2018.1.18
chardet==3.0.4
cssselect==1.0.3
Django==2.0.3
django-formtools==2.1
djangorestframework==3.7.7
eutils==0.3.2
idna==2.6
lxml==4.1.1
metapub==0.4.3.6
numpy==1.14.1
Pillow==5.0.0
pluggy==0.6.0
py==1.5.2
pytz==2018.3
requests==2.18.4
six==1.11.0
tabulate==0.8.2
tox==2.9.1
Unidecode==1.0.22
urllib3==1.22
virtualenv==15.1.0
